+++ src/controllers/RuleController.ts
@@ -1,20 +1,20 @@
 import { NextFunction, Request, Response } from 'express';
 import { Result, ValidationError, validationResult } from 'express-validator';
-import { HomiesWithIsTmpMemberResponseDto } from '../interfaces/rule/HomiesWithIsTmpMemberResponseDto';
-import { RuleCreateDto } from '../interfaces/rule/RuleCreateDto';
-import { RuleCreateInfoResponseDto } from '../interfaces/rule/RuleCreateInfoResponseDto';
-import { RuleHomeResponseDto } from '../interfaces/rule/RuleHomeResponseDto';
-import { RuleMyTodoResponseDto } from '../interfaces/rule/RuleMyTodoResponseDto';
-import { RuleReadInfoResponseDto } from '../interfaces/rule/RuleReadInfoResponseDto';
-import { RuleResponseDto } from '../interfaces/rule/RuleResponseDto';
-import { RulesByCategoryResponseDto } from '../interfaces/rule/RulesByCategoryResponseDto';
-import { RuleTodoCheckUpdateDto } from '../interfaces/rule/RuleTodoCheckUpdateDto';
-import { RuleTodoCheckUpdateResponseDto } from '../interfaces/rule/RuleTodoCheckUpdateResponseDto';
-import { RuleUpdateDto } from '../interfaces/rule/RuleUpdateDto';
-import { TmpRuleMembersUpdateResponseDto } from '../interfaces/rule/TmpRuleMembersUpdateResponseDto';
-import { RuleCategoryCreateDto } from '../interfaces/rulecategory/RuleCategoryCreateDto';
-import { RuleCategoryResponseDto } from '../interfaces/rulecategory/RuleCategoryResponseDto';
-import { RuleCategoryUpdateDto } from '../interfaces/rulecategory/RuleCategoryUpdateDto';
+import { RuleCreateDto } from '../interfaces/rule/request/RuleCreateDto';
+import { RuleTodoCheckUpdateDto } from '../interfaces/rule/request/RuleTodoCheckUpdateDto';
+import { RuleUpdateDto } from '../interfaces/rule/request/RuleUpdateDto';
+import { HomiesWithIsTmpMemberResponseDto } from '../interfaces/rule/response/HomiesWithIsTmpMemberResponseDto';
+import { RuleCreateInfoResponseDto } from '../interfaces/rule/response/RuleCreateInfoResponseDto';
+import { RuleHomeResponseDto } from '../interfaces/rule/response/RuleHomeResponseDto';
+import { RuleMyTodoResponseDto } from '../interfaces/rule/response/RuleMyTodoResponseDto';
+import { RuleReadInfoResponseDto } from '../interfaces/rule/response/RuleReadInfoResponseDto';
+import { RuleResponseDto } from '../interfaces/rule/response/RuleResponseDto';
+import { RulesByCategoryResponseDto } from '../interfaces/rule/response/RulesByCategoryResponseDto';
+import { RuleTodoCheckUpdateResponseDto } from '../interfaces/rule/response/RuleTodoCheckUpdateResponseDto';
+import { TmpRuleMembersUpdateResponseDto } from '../interfaces/rule/response/TmpRuleMembersUpdateResponseDto';
+import { RuleCategoryCreateDto } from '../interfaces/rulecategory/request/RuleCategoryCreateDto';
+import { RuleCategoryUpdateDto } from '../interfaces/rulecategory/request/RuleCategoryUpdateDto';
+import { RuleCategoryResponseDto } from '../interfaces/rulecategory/response/RuleCategoryResponseDto';
 import message from '../modules/responseMessage';
 import statusCode from '../modules/statusCode';
 import util from '../modules/util';
