+++ src/interfaces/common/request/LoginDto.ts
@@ -1,0 +1,5 @@
+export interface LoginDto {
+  email: string;
+  password: string;
+  fcmToken: string;
+}
