+++ src/interfaces/common/request/SignupDto.ts
@@ -1,0 +1,8 @@
+export interface SignupDto {
+  email: string;
+  password: string;
+  userName: string;
+  gender: string;
+  birthday?: Date;
+  fcmToken: string;
+}
