--- src/interfaces/event/EventCreateResponseDto.ts
@@ -1,6 +1,0 @@
-import { PostBaseResponseDto } from '../common/PostBaseResponseDto';
-import { EventCreateDto } from './EventCreateDto';
-
-export interface EventCreateResponseDto
-  extends EventCreateDto,
-    PostBaseResponseDto {}
