--- src/interfaces/event/EventUpdateResponseDto.ts
@@ -1,6 +1,0 @@
-import { PostBaseResponseDto } from '../common/PostBaseResponseDto';
-import { EventUpdateDto } from './EventUpdateDto';
-
-export interface EventUpdateResponseDto
-  extends EventUpdateDto,
-    PostBaseResponseDto {}
