+++ src/interfaces/event/response/EventCreateResponseDto.ts
@@ -1,0 +1,6 @@
+import { PostBaseResponseDto } from '../../common/response/PostBaseResponseDto';
+import { EventCreateDto } from '../request/EventCreateDto';
+
+export interface EventCreateResponseDto
+  extends EventCreateDto,
+    PostBaseResponseDto {}
