+++ src/interfaces/event/response/EventUpdateResponseDto.ts
@@ -1,0 +1,6 @@
+import { PostBaseResponseDto } from '../../common/response/PostBaseResponseDto';
+import { EventUpdateDto } from '../request/EventUpdateDto';
+
+export interface EventUpdateResponseDto
+  extends EventUpdateDto,
+    PostBaseResponseDto {}
