+++ src/interfaces/room/response/RoomJoinResponseDto.ts
