+++ src/interfaces/room/response/RoomResponseDto.ts
