+++ src/interfaces/rule/request/RuleUpdateDto.ts
