+++ src/interfaces/rule/response/HomiesWithIsTmpMemberResponseDto.ts
