+++ src/interfaces/rule/response/RuleCreateInfoResponseDto.ts
