+++ src/interfaces/rule/response/RuleHomeResponseDto.ts
