+++ src/interfaces/rule/response/RuleReadInfoResponseDto.ts
