+++ src/interfaces/rule/response/RuleResponseDto.ts
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose';
-import { RuleInfo } from './RuleInfo';
+import { RuleInfo } from '../RuleInfo';
 
 export interface RuleResponseDto extends RuleInfo {
   _id: mongoose.Types.ObjectId;
