+++ src/interfaces/rule/response/RulesByCategoryResponseDto.ts
