+++ src/interfaces/rulecategory/request/RuleCategoryUpdateDto.ts
