+++ src/interfaces/rulecategory/response/RuleCategoryResponseDto.ts
