--- src/interfaces/user/UserTypeTestResponseDto.ts
@@ -1,6 +1,0 @@
-import { PostBaseResponseDto } from '../common/PostBaseResponseDto';
-import { UserTypeTestDto } from './UserTypeTestDto';
-
-export interface UserTypeTestResponseDto
-  extends UserTypeTestDto,
-    PostBaseResponseDto {}
