+++ src/interfaces/user/response/UserNotificationStateUpdateDto.ts
