+++ src/interfaces/user/response/UserTypeTestResponseDto.ts
@@ -1,0 +1,6 @@
+import { PostBaseResponseDto } from '../../common/response/PostBaseResponseDto';
+import { UserTypeTestDto } from '../request/UserTypeTestDto';
+
+export interface UserTypeTestResponseDto
+  extends UserTypeTestDto,
+    PostBaseResponseDto {}
