+++ src/services/rule/RuleService.ts
@@ -1,15 +1,15 @@
 import dayjs from 'dayjs';
 import errorGenerator from '../../errors/errorGenerator';
-import { RuleCreateDto } from '../../interfaces/rule/RuleCreateDto';
-import { RuleResponseDto } from '../../interfaces/rule/RuleResponseDto';
-import { RuleTodoCheckUpdateDto } from '../../interfaces/rule/RuleTodoCheckUpdateDto';
-import { RuleTodoCheckUpdateResponseDto } from '../../interfaces/rule/RuleTodoCheckUpdateResponseDto';
-import { RuleUpdateDto } from '../../interfaces/rule/RuleUpdateDto';
-import { TmpRuleMembersUpdateDto } from '../../interfaces/rule/TmpRuleMembersUpdateDto';
-import { TmpRuleMembersUpdateResponseDto } from '../../interfaces/rule/TmpRuleMembersUpdateResponseDto';
-import { RuleCategoryCreateDto } from '../../interfaces/rulecategory/RuleCategoryCreateDto';
-import { RuleCategoryResponseDto } from '../../interfaces/rulecategory/RuleCategoryResponseDto';
-import { RuleCategoryUpdateDto } from '../../interfaces/rulecategory/RuleCategoryUpdateDto';
+import { RuleCreateDto } from '../../interfaces/rule/request/RuleCreateDto';
+import { RuleTodoCheckUpdateDto } from '../../interfaces/rule/request/RuleTodoCheckUpdateDto';
+import { RuleUpdateDto } from '../../interfaces/rule/request/RuleUpdateDto';
+import { TmpRuleMembersUpdateDto } from '../../interfaces/rule/request/TmpRuleMembersUpdateDto';
+import { RuleResponseDto } from '../../interfaces/rule/response/RuleResponseDto';
+import { RuleTodoCheckUpdateResponseDto } from '../../interfaces/rule/response/RuleTodoCheckUpdateResponseDto';
+import { TmpRuleMembersUpdateResponseDto } from '../../interfaces/rule/response/TmpRuleMembersUpdateResponseDto';
+import { RuleCategoryCreateDto } from '../../interfaces/rulecategory/request/RuleCategoryCreateDto';
+import { RuleCategoryUpdateDto } from '../../interfaces/rulecategory/request/RuleCategoryUpdateDto';
+import { RuleCategoryResponseDto } from '../../interfaces/rulecategory/response/RuleCategoryResponseDto';
 import Check from '../../models/Check';
 import Rule from '../../models/Rule';
 import RuleCategory from '../../models/RuleCategory';
