+++ test/RoomService.spec.ts
@@ -1,17 +1,17 @@
 import assert from 'assert';
 import { afterEach } from 'mocha';
 import config from '../src/config';
-import { SignupDto } from '../src/interfaces/auth/SignupDto';
-import { PostBaseResponseDto } from '../src/interfaces/common/PostBaseResponseDto';
-import { RoomJoinDto } from '../src/interfaces/room/RoomJoinDto';
-import { RoomResponseDto } from '../src/interfaces/room/RoomResponseDto';
+import { SignupDto } from '../src/interfaces/auth/request/SignupDto';
+import { PostBaseResponseDto } from '../src/interfaces/common/response/PostBaseResponseDto';
+import { RoomJoinDto } from '../src/interfaces/room/request/RoomJoinDto';
+import { RoomResponseDto } from '../src/interfaces/room/response/RoomResponseDto';
 import connectDB from '../src/loaders/db';
 import Event from '../src/models/Event';
 import Room from '../src/models/Room';
 import RuleCategory from '../src/models/RuleCategory';
 import User from '../src/models/User';
-import RoomService from '../src/services/RoomService';
-import UserService from '../src/services/UserService';
+import RoomService from '../src/services/room/RoomService';
+import UserService from '../src/services/user/UserService';
 
 describe('RoomService Tests', () => {
   if (config.env !== 'test') {
